# map: optimize memory allocation when building the map