Before:
```
2025/03/13 15:58:08 Generated 100x200 Arcanus map in 53.391247ms
2025/03/13 15:58:08 Generated 100x200 Arcanus map in 54.024643ms
2025/03/13 15:58:08 Generated 100x200 Arcanus map in 56.269528ms
```
After:
```
2025/03/13 15:57:53 Generated 100x200 Arcanus map in 44.805507ms
2025/03/13 15:57:53 Generated 100x200 Arcanus map in 38.5493ms
2025/03/13 15:57:54 Generated 100x200 Arcanus map in 41.392034ms
```